[luci/pass] Relocate has_dynamic_shape to helper/Shape
This will relocal has_dynamic_shape method to helpers/Shape module.

ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
